requirements.txt
gunicorn
flask
opencv-python
Pillow
moviepy
requests
python-dotenv